+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog
 ## 2.0.1
 - Migration to dart:js_interop & package:web/web.dart
 - Migration to Flutter 3.22
+- Add a description for each tx to sign to explain the goal of the tx
 
 ## 2.0.0
 - Add models for all RPC **Requests** and **Responses**